Move invite-member request to POST api
